refactor(cache-manager): move memory/redis store files into stores/ subfolder
Move memory-store.ts, redis-store.ts and their tests into a dedicated
stores/ subfolder to separate pluggable store backends from core module
files. Update all internal import paths accordingly while preserving the
identical public API re-export surface in index.ts.

Refs #858
