+++ packages/cache-manager/src/index.ts
@@ -1,8 +1,8 @@
 export * from './decorators.js';
 export * from './interceptor.js';
-export * from './memory-store.js';
+export * from './stores/memory-store.js';
 export * from './module.js';
-export * from './redis-store.js';
+export * from './stores/redis-store.js';
 export * from './service.js';
 export * from './status.js';
 export * from './tokens.js';
