+++ packages/cache-manager/src/module.ts
@@ -3,8 +3,8 @@ import { defineModule, type ModuleType } from '@konekti/runtime';
 import { RUNTIME_CONTAINER } from '@konekti/runtime/internal';
 
 import { CacheInterceptor } from './interceptor.js';
-import { MemoryStore } from './memory-store.js';
-import { RedisStore } from './redis-store.js';
+import { MemoryStore } from './stores/memory-store.js';
+import { RedisStore } from './stores/redis-store.js';
 import { CacheService } from './service.js';
 import { CACHE_OPTIONS, CACHE_STORE } from './tokens.js';
 import type { CacheModuleOptions, NormalizedCacheModuleOptions, RedisCompatibleClient } from './types.js';
